LPS-75043 Renaming UpgradeProcess_7_0_4 to UpgradeProcess_7_0_5 since this class has not been backported for 7.0.4 GA5
