+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_5.java
@@ -16,26 +16,26 @@ package com.liferay.portal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeCompany;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeContact;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeEmailAddress;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeExpando;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeGroup;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeMBMailingList;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeModules;
-import com.liferay.portal.upgrade.v7_0_4.UpgradePortalPreferences;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeUser;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeVirtualHost;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeCompany;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeContact;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeEmailAddress;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeExpando;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeGroup;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeMBMailingList;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeModules;
+import com.liferay.portal.upgrade.v7_0_5.UpgradePortalPreferences;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeUser;
+import com.liferay.portal.upgrade.v7_0_5.UpgradeVirtualHost;
 
 /**
  * @author Roberto Díaz
  * @author Samuel Ziemer
  */
-public class UpgradeProcess_7_0_4 extends UpgradeProcess {
+public class UpgradeProcess_7_0_5 extends UpgradeProcess {
 
 	@Override
 	public int getThreshold() {
-		return ReleaseInfo.RELEASE_7_0_4_BUILD_NUMBER;
+		return ReleaseInfo.RELEASE_7_0_5_BUILD_NUMBER;
 	}
 
 	@Override
