+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_5/UpgradeCompany.java
@@ -12,22 +12,20 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_0_4;
+package com.liferay.portal.upgrade.v7_0_5;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.upgrade.v7_0_4.util.MBMailingListTable;
+import com.liferay.portal.upgrade.v7_0_5.util.CompanyTable;
 
 /**
- * @author Ugurcan Cetin
+ * @author Jose A. Jimenez Campoy
  */
-public class UpgradeMBMailingList extends UpgradeProcess {
+public class UpgradeCompany extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		alter(
-			MBMailingListTable.class,
-			new AlterColumnType("emailAddress", "VARCHAR(254) null"),
-			new AlterColumnType("outEmailAddress", "VARCHAR(254) null"));
+			CompanyTable.class, new AlterColumnType("mx", "VARCHAR(200) null"));
 	}
 
 }
