+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_5/UpgradeContact.java
@@ -12,20 +12,21 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_0_4;
+package com.liferay.portal.upgrade.v7_0_5;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.upgrade.v7_0_4.util.CompanyTable;
+import com.liferay.portal.upgrade.v7_0_5.util.ContactTable;
 
 /**
- * @author Jose A. Jimenez Campoy
+ * @author Ugurcan Cetin
  */
-public class UpgradeCompany extends UpgradeProcess {
+public class UpgradeContact extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		alter(
-			CompanyTable.class, new AlterColumnType("mx", "VARCHAR(200) null"));
+			ContactTable.class,
+			new AlterColumnType("emailAddress", "VARCHAR(254) null"));
 	}
 
 }
