+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_5/util/VirtualHostTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_0_4.util;
+package com.liferay.portal.upgrade.v7_0_5.util;
 
 import java.sql.Types;
